+++ tests/Agent/UnitTests/Core.UnitTest/Transactions/TransactionTests.cs
@@ -537,4 +537,36 @@ public class TransactionTests
         // Assert
         Assert.That(_transaction.TransactionMetadata.UserAndRequestAttributes.Count, Is.EqualTo(0));
     }
+
+    [Test]
+    public void AddFaasAttribute_SetAttributeInTransactionMetadata()
+    {
+        // Arrange
+        var key = "TestAttribute";
+        var value = "TestValue";
+
+        // Act
+        _transaction.AddFaasAttribute(key, value);
+
+        // Assert
+        var allAttributeValuesDic = _transaction.TransactionMetadata.UserAndRequestAttributes.GetAllAttributeValuesDic();
+
+        var attributeValue = allAttributeValuesDic[key];
+        Assert.That(attributeValue, Is.EqualTo(value));
+    }
+
+    [TestCase("   ")]
+    [TestCase("")]
+    [TestCase(null)]
+    public void AddFaasAttribute_DoesNotSetAttribute_WhenKeyIsBad(string key)
+    {
+        // Arrange
+        var value = "TestValue";
+
+        // Act
+        _transaction.AddFaasAttribute(key, value);
+
+        // Assert
+        Assert.That(_transaction.TransactionMetadata.UserAndRequestAttributes.Count, Is.EqualTo(0));
+    }
 }